glmark2: add gbm offscreen backend
glmark2 supports offscreen runs and even gbm already in the case of the
drm backend. However, in all cases even with the --off-screen flag, it
goes through display or a window system initialization which requires
connected hardware or additional environment setup to satisfy that
initialization (connected monitor or Xorg or a Wayland compositor
running).

This commit adds a simple gbm backend which is purely offscreen and
requires no display or window system setup.
The advantage is being able to run on systems with no display-related
requirements, and providing GPU benchmark results completely independent
of displays.
It is of course not possible to visually see the results though, as is
already the case with other offscreen runs.

This is a very similar runtime flow as the drm backend with
--off-screen, but does not attempt a connection with DRM to query
parameters such as valid connector or the display resolution.
Additionally, compared to the drm backend this allows any surface size
to be specified, not tied to display resolutions.

The benchmark results are equivalent to a run with drm backend and
--off-screen for the same surface size, so compatible with current
glmark2 scores.

This backend can be helpful in, for instance: development systems,
automated test runs, CI environments, or runs on headless remote
systems.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>